edge-device-stack/edge-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-agent"
version = "0.1.0"
description = "Edge agent for KIN-GOLIATH farm gateway"
authors = [{name = "IoT Edge Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "paho-mqtt>=1.6",
    "pyserial>=3.5",
    "cbor2>=5.4",
    "msgpack>=1.0",
    "apscheduler>=3.10",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "aiofiles>=23.2",
    "tenacity>=8.2",
    "orjson>=3.10",
    "jinja2>=3.1",
    "httpx>=0.27",
    "brotli>=1.1",
    "uvloop>=0.19; platform_system != 'Windows'",
    "ruff>=0.5",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "pytest-httpx>=0.30",
    "freezegun>=1.5",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "ASYNC", "S", "BLE"]
ignore = ["S101"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["edge-agent/tests"]
asyncio_mode = "auto"
